requirements.txt
aiohttp
fastapi
jinja2
python-dotenv
uvicorn[standard]
